+++ frontend/app/.server/routes/helpers/renew-route-helpers.ts
@@ -2,6 +2,7 @@ import type { Session } from '@remix-run/node';
 import { redirectDocument } from '@remix-run/node';
 import type { Params } from '@remix-run/react';
 
+import { merge } from 'moderndash';
 import { z } from 'zod';
 
 import type { ClientApplicationDto } from '~/.server/domain/dtos';
@@ -60,17 +61,20 @@ export interface RenewState {
   readonly hasAddressChanged?: boolean;
   readonly isHomeAddressSameAsMailingAddress?: boolean;
   readonly addressInformation?: {
-    copyMailingAddress: boolean;
+    // All properties made optional since we are potentially saving mailing and home from two different pages (update-mailing-address/update-home-address).
+    // Refactoring solution: split addressInformation into two: homeAddress, and mailingAddress.
+    // Following Figma, after implementation, 'copyMailingAddress' might be useless.
+    copyMailingAddress?: boolean;
     homeAddress?: string;
     homeApartment?: string;
     homeCity?: string;
     homeCountry?: string;
     homePostalCode?: string;
     homeProvince?: string;
-    mailingAddress: string;
+    mailingAddress?: string;
     mailingApartment?: string;
-    mailingCity: string;
-    mailingCountry: string;
+    mailingCity?: string;
+    mailingCountry?: string;
     mailingPostalCode?: string;
     mailingProvince?: string;
   };
@@ -181,10 +185,9 @@ export function saveRenewState({ params, session, state }: SaveStateArgs) {
   const log = getLogger('renew-route-helpers.server/saveRenewState');
   const currentState = loadRenewState({ params, session });
 
-  const newState = {
-    ...currentState,
-    ...state,
-  } satisfies RenewState;
+  // https://moderndash.io/docs/merge
+  // Merges two objects without overwriting any unmentioned properties
+  const newState = merge({}, currentState, state) satisfies RenewState;
 
   const sessionName = getSessionName(currentState.id);
   session.set(sessionName, newState);
