+++ frontend/app/routes/public/renew/$id/ita/confirmation.tsx
@@ -269,8 +269,8 @@ export default function RenewFlowConfirm() {
               {mailingAddressInfo ? (
                 <Address
                   address={{
-                    address: mailingAddressInfo.address,
-                    city: mailingAddressInfo.city,
+                    address: mailingAddressInfo.address ?? '',
+                    city: mailingAddressInfo.city ?? '',
                     provinceState: mailingAddressInfo.province,
                     postalZipCode: mailingAddressInfo.postalCode,
                     country: mailingAddressInfo.country,
